Rework category indexes to allow subheadings.
Each asciidoc inddex now invokes an erb partial to list a set of pages.
Pages can be tagged with a comma separated :page-tag:
Nothing ensures that there aren't absent pages.

Asciidoc needs to see a section heading to generate a preamble, and the
css spacing is also easier if one exists, but if there's no subheadings
then it looks weird to add one.  So instead, the section heading is
tagged with some css to make the text disappear and have zero height,
but its margins remain.