Optimize local workflow
Rewrite serve.py to do bespoke file watching and command running, so
that we can make middleman only renerate a changed html file for the
common case of document editing.  This brings the edit->refresh cycle to
8 seconds.

Also disable asset hashing for local builds.  This brings the
edit->refresh cycle down to 4 seconds.